feat: stripe monthly renew handling
- now after each monthly renew the token will get reset for usage.
- if payment fails then the plan deactivates until the payment is completed from billing portal
- if payment success then and only then the counter is reset.

+++ src/app/api/webhook/route.tsx
@@ -25,8 +25,10 @@ export async function POST(request: NextRequest) {
       const session: Stripe.Checkout.Session = event.data.object;
       const userId = session.metadata?.user_id as string;
       const priceId = session.metadata?.price_id as string;
-      const maxUsageUnit = parseInt(session.metadata?.max_usage_limit as string);
-      
+      const maxUsageUnit = parseInt(
+        session.metadata?.max_usage_limit as string
+      );
+
       await SubscriptionRepo.assignSubscription(
         userId,
         session.customer as string,
@@ -34,23 +36,72 @@ export async function POST(request: NextRequest) {
         session.subscription as string,
         maxUsageUnit
       );
-    }
-
-    if (event.type === "customer.subscription.updated") {
+    } else if (event.type === "customer.subscription.updated") {
       const subscription: Stripe.Subscription = event.data.object;
-      await SubscriptionRepo.updateExpires(
-        subscription.id,
-        subscription.cancel_at as number
-      );
-    }
 
-    if (event.type === "customer.subscription.deleted") {
+      if (
+        await SubscriptionRepo.findSubscriptionByStripeSubscriptionId(
+          subscription.id
+        )
+      ) {
+        await SubscriptionRepo.updateExpires(
+          subscription.id,
+          subscription.cancel_at as number
+        );
+      }
+    } else if (event.type === "invoice.paid") {
+      const invoice = event.data.object as Stripe.Invoice;
+      if (
+        await SubscriptionRepo.findSubscriptionByStripeSubscriptionId(
+          invoice.subscription as string
+        )
+      ) {
+        // reset the token usage
+        await SubscriptionRepo.markAsSuccessfullyRenewed(
+          invoice.subscription as string
+        );
+      } else {
+        if (invoice.billing_reason !== "subscription_create") {
+          throw new Error("Subscription not found in database!");
+        }
+      }
+    } else if (event.type === "invoice.payment_failed") {
+      const invoice = event.data.object as Stripe.Invoice;
+
+      // reset the token usage
+      const subscription =
+        await SubscriptionRepo.findSubscriptionByStripeSubscriptionId(
+          invoice.subscription as string
+        );
+      if (subscription) {
+        // if the subscription is active, we can stop it by marking it as failed and email the user
+        if (!subscription.stripeInvoiceFailed) {
+          // TODO: email the user about the payment failure and tell them to update their payment method in billing portal
+
+          await SubscriptionRepo.markAsFailedToRenew(
+            invoice.subscription as string
+          );
+        }
+      } else {
+        if (invoice.billing_reason !== "subscription_create") {
+          throw new Error("Subscription not found in database!");
+        }
+      }
+    } else if (event.type === "customer.subscription.deleted") {
       const subscription = event.data.object;
 
-      await SubscriptionRepo.deleteSubscription(
-        subscription.id,
-        subscription.cancel_at as number
-      );
+      if (
+        await SubscriptionRepo.findSubscriptionByStripeSubscriptionId(
+          subscription.id
+        )
+      ) {
+        await SubscriptionRepo.deleteSubscription(
+          subscription.id,
+          subscription.cancel_at as number
+        );
+      }
+    } else {
+      throw new Error(`Invalid event type: ${event.type}!`);
     }
 
     return NextResponse.json({ message: "success" });
